requirements.txt
streamlit>=1.28.0
ollama>=0.1.7
Pillow>=10.0.0
requests>=2.31.0
langchain>=0.0.350
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
python-docx>=1.1.0